pyproject.toml
[project]
name = "local-llm-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "diskcache==5.6.3",
    "jinja2==3.1.6",
    "llama-cpp-python==0.3.16",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "numpy==2.4.2",                                                                                                                                                                                                      
    "pygments==2.19.2",
    "rich==14.3.2",
    "typing-extensions==4.15.0"
]
